move java datalog lib
