+++ libdatalog/src/main/java/edu/wpi/first/datalog/BooleanArrayLogEntry.java
