+++ libdatalog/src/main/java/edu/wpi/first/datalog/BooleanLogEntry.java
