+++ libdatalog/src/main/java/edu/wpi/first/datalog/DataLogIterator.java
