+++ libdatalog/src/main/java/edu/wpi/first/datalog/DoubleArrayLogEntry.java
