+++ libdatalog/src/main/java/edu/wpi/first/datalog/DoubleLogEntry.java
