+++ libdatalog/src/main/java/edu/wpi/first/datalog/FloatLogEntry.java
