+++ libdatalog/src/main/java/edu/wpi/first/datalog/IntegerLogEntry.java
