+++ libdatalog/src/main/java/edu/wpi/first/datalog/RawLogEntry.java
