+++ libdatalog/src/main/java/edu/wpi/first/datalog/StringLogEntry.java
