+++ libdatalog/src/main/java/edu/wpi/first/datalog/StructLogEntry.java
